enable HMR + resolve conflict
