+++ superset/assets/src/addSlice/index.jsx
@@ -1,14 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { appSetup } from '../common';
-import AddSliceContainer from './AddSliceContainer';
-
-appSetup();
-
-const addSliceContainer = document.getElementById('js-add-slice-container');
-const bootstrapData = JSON.parse(addSliceContainer.getAttribute('data-bootstrap'));
+import App from './App';
 
 ReactDOM.render(
-  <AddSliceContainer datasources={bootstrapData.datasources} />,
-  addSliceContainer,
+  <App />,
+  document.getElementById('js-add-slice-container'),
 );
